Remove babel-node command from production server script
Per Babel docs, the `babel-node` command is not meant for production use:
https://babeljs.io/docs/usage/cli/#babel-node

```
Not meant for production use
You should not be using babel-node in production. It is unnecessarily heavy, with high memory usage due to the cache being stored in memory. You will also always experience a startup performance penalty as the entire app needs to be compiled on the fly.
```